moved package and subpackages to their __init__; updated manifest.json for templated setup.py; moved full loop assignment to init from run_in_eventloop
