+++ setup.py
@@ -1,31 +1,32 @@
+import os
 import json
 
-from setuptools import setup, find_packages
+from setuptools import find_packages, setup
 
-with open('README.md', 'r') as fin:
+with open(os.path.join(os.path.dirname(__file__), 'README.md')) as fin:
     README = fin.read()
 
-with open('manifest.json', 'r') as fin:
+with open(os.path.join(os.path.dirname(__file__), 'manifest.json')) as fin:
     manifest = json.load(fin)
 
-setup (
-    name = manifest.get('name', 'manifest.name'),
-    version = manifest.get('version', '0.0.0'),
-    author = manifest.get('author', 'manifest.author'),
-    author_email = manifest.get('author_email', 'manifest.author.email'),
-    description = manifest.get('description', 'manifest.description'),
+
+setup(
+    name = manifest.get('name', 'default.manifest.name'),
+    version =  manifest.get('version', '0.0.0'),
+    packages = find_packages(
+        include=manifest.get('package_include', []), 
+        exclude=manifest.get('package_exclude', [])
+    ),
+    package_data = manifest.get('package_data', None),
+    include_package_data = manifest.get('include_package_data', True),
+    description =  manifest.get('description', 'default.manifest.description'),
     long_description = README,
-    long_description_content_type = 'text/markdown',
-    url = 'https://www.github.com/tannerburns/vast',
-    packages = find_packages(),
-    include_package_data = True,
-    install_requires = [
-        'requests',
-        'tqdm',
-        'colored'
-    ],
-    classifiers = [
-        'Programming Language :: Python :: 3',
-        'Operating System :: OS Independent',
-    ],
+    url = manifest.get('url', 'default.manifest.url'),
+    author = manifest.get('author', 'default.manifest.author'),
+    author_email = manifest.get('author_email', 'default.manifest.author_email'),
+    install_requires = manifest.get('requirements', []),
+    classifiers =  manifest.get('classifiers', []),
+    entry_points = manifest.get('entry_points', {}),
+    scripts = manifest.get('scripts', None),
+    keywords = manifest.get('keywords', [])
 )
