+++ vast/decorators/__init__.py
@@ -1,1 +1,13 @@
-from .vastdecorators import vast_loop
+from .. import Vast
+
+class vast_loop(Vast):
+
+    def __init__(self, workers: int= 16):
+        super().__init__(workers=workers)
+    
+    def __call__(self, fn):
+        
+        def vloop(listofArgs: list, **kwargs):
+            return self.run_in_eventloop(fn, listofArgs, **kwargs)
+        
+        return vloop
